♻️[#34] Refactor: types 분리 및 selectedIconColor로 변경
