+++ packages/design-system/src/components/RadioGroup/RadioGroup.tsx
@@ -3,56 +3,11 @@ import { cloneElement, isValidElement, type ReactElement } from 'react';
 import React from 'react';
 import { twMerge } from 'tailwind-merge';
 
+import type { RadioContextType, RadioProps } from './types';
+
 //////////////////////////////////////////
 // Context 타입 정의
 //////////////////////////////////////////
-
-/**
- * RadioGroup 컨텍스트의 타입 정의입니다.
- *
- * @property {string} [title] - 라디오 그룹 상단의 제목 (선택 사항)
- * @property {string} [titleClassName] - 제목의 커스텀 클래스명
- * @property {string} [radioGroupClassName] - 라디오 그룹(wrapper)의 커스텀 클래스명
- * @property {string | number} [selectedValue] - 현재 선택된 라디오의 값
- * @property {(value: string | number) => void} onSelect - 라디오 선택 시 호출되는 콜백
- * @property {React.ReactNode} [children] - 하위 라디오 버튼 요소들
- * @property {string} [selectedColor] - 선택된 라디오 버튼의 아이콘 색상 (기본값: 'white')
- * @property {string} [selectedClassName] - 선택된 라디오 버튼에 적용할 커스텀 클래스
- * @property {string} [unSelectedClassName] - 비선택된 라디오 버튼에 적용할 커스텀 클래스
- */
-
-interface RadioContextType {
-  title?: string;
-  titleClassName?: string;
-  radioGroupClassName?: string;
-  selectedValue?: string | number;
-  onSelect?: (value: string | number) => void;
-  children?: React.ReactNode;
-  selectedColor?: string;
-  selectedClassName?: string;
-  unSelectedClassName?: string;
-}
-
-/**
- * Radio 버튼 컴포넌트의 props 정의입니다.
- *
- * @property {string | number} value - 고유 식별자. 이 값이 selectedValue와 일치하면 선택 상태로 간주됩니다.
- * @property {string} [name='radioGroup'] - input의 name 속성 (기본: 'radioGroup')
- * @property {React.ReactNode} [children] - 버튼 안에 들어갈 내용 (텍스트, 아이콘 등)
- * @property {string} [selectedColor] - 선택된 경우 아이콘에 적용할 색상
- * @property {string} [selectedClassName] - 선택된 상태에서의 커스텀 클래스
- * @property {string} [unSelectedClassName] - 비선택 상태에서의 커스텀 클래스
- *
- */
-
-interface RadioProps extends Omit<React.InputHTMLAttributes<HTMLInputElement>, 'value' | 'onChange'> {
-  value: number | string;
-  name?: string;
-  children?: React.ReactNode;
-  selectedColor?: string;
-  selectedClassName?: string;
-  unSelectedClassName?: string;
-}
 
 //////////////////////////////////////////
 // Context 및 훅
@@ -98,7 +53,7 @@ function useRadioContext() {
            onSelect={setSelectedCategory1}
            unSelectedClassName='bg-purple-200 text-gray-500'
            selectedClassName='bg-red-400 text-green-300'
-           selectedColor='red'
+           selectedIconColor='red'
          >
            <div className='flex gap-12'>
              <RadioGroup.Radio value='Art' className='flex gap-8'>
@@ -124,7 +79,7 @@ export default function RadioGroup({
   selectedValue,
   onSelect,
   children,
-  selectedColor,
+  selectedIconColor,
   selectedClassName,
   unSelectedClassName,
 }: RadioContextType) {
@@ -136,7 +91,7 @@ export default function RadioGroup({
         radioGroupClassName,
         selectedValue,
         onSelect,
-        selectedColor,
+        selectedIconColor,
         selectedClassName,
         unSelectedClassName,
       }}
@@ -157,12 +112,12 @@ RadioGroup.Radio = function Radio({ value, children, className = '', name = 'rad
   const {
     selectedValue,
     onSelect,
-    selectedColor: groupSelectedColor,
+    selectedIconColor: groupSelectedIconColor,
     selectedClassName: groupSelectedClassName,
     unSelectedClassName: groupUnselectedClassName,
   } = useRadioContext();
 
-  const resolvedSelectedColor = props.selectedColor ?? groupSelectedColor ?? 'white';
+  const resolvedSelectedColor = props.selectedIconColor ?? groupSelectedIconColor ?? 'white';
   const resolvedSelectedClassName = props.selectedClassName ?? groupSelectedClassName ?? '';
   const resolvedUnselectedClassName = props.unSelectedClassName ?? groupUnselectedClassName ?? '';
