+++ packages/design-system/src/components/index.ts
@@ -3,6 +3,6 @@ export * from './icons';
 export * from './input';
 export * from './logos';
 export { default as Pagination } from './Pagination';
-export { default as RadioGroup } from './RadioGroup';
+export { default as RadioGroup } from './RadioGroup/RadioGroup';
 export * from '@/components/calendar/index';
 export * from '@/components/dropdown/index';
